Added new abstract class ddp_generator.py that handles most of autogen.
This class is meant to crate a very simply way for new languages to be autogenerated. It handles all of the logic for taking a jinja template and turning it into the output necessary for protoc and relies only a few simple overrides to figure out how to name output files.